Saving

+++ render.py
@@ -1,11 +1,15 @@
 #!/opt/python/bin/python
-import argparse, time, tempfile, subprocess, os, shutil
+import argparse, time, tempfile, subprocess, os, shutil, math
 
 def die(err):
     print "Error: {}".format(err)
     exit(1)
 
-MAX_TIME_PER_JOB = 60 * 20 # 20 min in default queue
+MAX_TIME_PER_JOB = 60 * 60 * 2 # 2 hours in kayvon queue
+NODES = []
+
+for i in range(5, 31, 2):
+    NODES.append('compute-0-{}'.format(i))
 
 def main():
     parser = argparse.ArgumentParser(
@@ -15,48 +19,49 @@ def main():
     parser.add_argument('scene', help='Path of scene to render relative to project')
     parser.add_argument('start', help='Starting frame to render', type=int)
     parser.add_argument('end', help='Ending frame to render', type=int)
-    parser.add_argument('-t', '--time_per_frame', type=int, default=1,
-                        help='Upper bound on render time per frame in minutes')
-    parser.add_argument('-a', '--args', default='',
-                        help='Additional arguments to pass to the Maya renderer.')
+    parser.add_argument('-s', '--skip', type=int, default=1, help='Render every X frames')
+    parser.add_argument('-t', '--time_per_frame', type=int, default=60,
+                        help='Upper bound on render time per frame in seconds')
+    parser.add_argument('-m', '--max_time_per_job', type=int, default=MAX_TIME_PER_JOB,
+                        help='Maximum time in seconds allowed for each job')
     parser.add_argument('-o', '--output', default='/home/wcrichto/maya-output',
                         help='Directory to output rendered frames to')
+    parser.add_argument('-a', '--args', default='',
+                        help='Additional arguments to pass to the Maya renderer.')
     args = parser.parse_args()
 
     if args.start > args.end:
         die("Start frame must be less than or equal to end frame")
 
-    total_frames = args.end - args.start + 1
-    total_time = total_frames * args.time_per_frame * 60
-    jobs = max(total_time / MAX_TIME_PER_JOB, 1)
-    frames_per_job = total_frames / jobs
-    extra_frames = total_frames % jobs
-    job_walltime = time.strftime('%H:%M:%S', time.gmtime(MAX_TIME_PER_JOB))
-
-    # print "Frames %d, jobs %d, frames per job: %d" % (total_frames, jobs, frames_per_job)
+    args.time_per_frame /= args.skip
+    total_frames = (args.end - args.start + 1)
+    frames_per_job = args.max_time_per_job / args.time_per_frame
+    jobs = int(math.ceil(float(total_frames) / frames_per_job))
+    extra_frames = total_frames % frames_per_job
+    job_walltime = time.strftime('%H:%M:%S', time.gmtime(frames_per_job * args.time_per_frame))
 
     with open('template.job', 'r') as f: job_template = f.read()
-    shutil.rmtree('jobs')
+    if os.path.isdir('jobs'): shutil.rmtree('jobs')
     os.makedirs('jobs')
 
     for i in xrange(jobs):
         job_start = i * frames_per_job + args.start
-        job_end = job_start + frames_per_job - 1
-        if i == jobs - 1: job_end += extra_frames
+        if i == jobs - 1: job_end = job_start + extra_frames - 1
+        else: job_end = job_start + frames_per_job - 1
         format_args = {
             'walltime': job_walltime,
             'start': job_start,
             'end': job_end,
-            'index': i,
+            'index': '{:02d}'.format(i),
             'args': args.args,
             'project': args.project,
             'scene': args.scene,
-            'output': args.output
+            'output': args.output,
+            'node': NODES[i % len(NODES)],
+            'skip': args.skip
             }
-        with open('jobs/maya_render_{}.job'.format(i), 'wb') as job_file:
+        with open('jobs/maya_render_{:02d}.job'.format(i), 'wb') as job_file:
             job_file.write(job_template.format(**format_args))
-            #subprocess.check_call('qsub {}'.format(job_file.name), shell=True)
-
 
 if __name__ == "__main__":
     main()
